requirements.txt
pytesseract
pillow
opencv-python
pdfplumber
PyMuPDF
